Changed all the floats to long double.
The factorial operation did not have enough precision. Even now, it
isn't great, but it is much better.

+++ include/factorial.h
@@ -1,6 +1,6 @@
 #ifndef _FACTORIAL_
 #define _FACTORIAL_
 
-float factorial(int x);
+long double factorial(int x);
 
 #endif
